# repo/backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Philatelic Museum Operations Suite API
  version: 1.0.0
  description: |
    Public HTTP API for museum operations.

    Authentication model:
    - Session cookie `museum_sid` (set on `POST /auth/login`)
    - `X-CSRF-Token` required on mutating requests (`POST|PUT|PATCH|DELETE`) except `POST /auth/login`
    - `X-Step-Up-Token` required on sensitive actions after `POST /auth/step-up`

    Step-up actions:
    - `GRAPH_PUBLISH`
    - `ROUTE_RULE_CHANGE`
    - `JOB_APPROVE`
    - `JOB_APPEAL_DECIDE`
    - `EXPORT_CREATE`
    - `ADMIN_CONFIG_UPDATE`
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - SessionCookie: []
paths:
...
  /catalog/search:
    get:
      tags: [Catalog]
      summary: Catalog search
      description: Publicly accessible search with caching and fuzzy relevance.
      security: []
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: sort
          in: query
          schema:
            type: string
            enum: [relevance:desc, title:asc, period:asc]
        - name: filter[category]
          in: query
          schema: { type: string }
        - name: filter[tags]
          in: query
          schema: { type: string }
        - name: filter[periodId]
          in: query
          schema: { type: string }
        - name: filter[seriesId]
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              example:
                data:
                  - id: 65f0f7d1d5f5f8612f06b111
                    title: Blue Airmail
                    catalogNumber: CAT-123
                    artist: I. Kline
                    series: Airmail
                    country: USA
                    period: 1930s
                    tags: [airmail, blue]
                pagination:
                  page: 1
                  pageSize: 20
                  total: 1
                  totalPages: 1
                meta:
                  cache: MISS
        '400': { $ref: '#/components/responses/ValidationError' }
  /catalog/autocomplete:
    get:
      tags: [Catalog]
      summary: Autocomplete suggestions
      security: []
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 64 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 15, default: 8 }
      responses:
        '200': { description: Suggestions }
        '400': { $ref: '#/components/responses/ValidationError' }
  /catalog/hot-keywords:
    get:
      tags: [Catalog]
      summary: Active curated hot keywords
      security: []
      responses:
        '200': { description: Keywords }
    post:
      tags: [Catalog]
      summary: Create hot keyword
      description: Requires `CATALOG_CURATION` (Administrator, Curator).
      security:
        - SessionCookie: []
          CsrfHeader: []
      responses:
        '201': { description: Created }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /catalog/hot-keywords/{keywordId}:
    patch:
      tags: [Catalog]
      summary: Update hot keyword
      description: Requires `CATALOG_CURATION`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: keywordId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Updated }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Catalog]
      summary: Retire hot keyword
      description: Requires `CATALOG_CURATION`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: keywordId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Retired }
        '404': { $ref: '#/components/responses/NotFound' }
  /catalog/items:
    post:
      tags: [Catalog]
      summary: Create catalog item
      description: Requires `CATALOG_CURATION`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      responses:
        '201': { description: Created }
        '400': { $ref: '#/components/responses/ValidationError' }
  /catalog/items/{itemId}:
    patch:
      tags: [Catalog]
      summary: Update catalog item
      description: Requires `CATALOG_CURATION`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Updated }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Catalog]
      summary: Archive catalog item
      description: Requires `CATALOG_CURATION`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Archived }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: museum_sid
      description: Session cookie set by `POST /auth/login`.
    CsrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on mutating requests except `POST /auth/login`.
    StepUpHeader:
      type: apiKey
      in: header
      name: X-Step-Up-Token
      description: One-time action-bound step-up token for sensitive operations.
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              oneOf:
                - type: array
                  items: { type: object }
                - type: object
            requestId: { type: string }
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Forbidden:
      description: Insufficient role permission or object-level access denied
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...